% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_question}
\alias{validate_question}
\title{Validate responses to questions
Validate responses to questions}
\usage{
validate_question(question, answers)
}
\description{
Validate responses to questions
Validate responses to questions
}
